Split LobattoInterpolations into header and cpp
