+++ test/mechanics/interpolation/CMakeLists.txt
@@ -28,20 +28,25 @@ add_unit_test(InterpolationPyramidLinear)
 add_unit_test(InterpolationTrussLobatto
     math/Legendre.cpp)
 add_unit_test(InterpolationQuadLobatto
+    mechanics/interpolation/InterpolationTrussLobatto.cpp
     math/Legendre.cpp)
 add_unit_test(InterpolationBrickLobatto
+    mechanics/interpolation/InterpolationTrussLobatto.cpp
     math/Legendre.cpp)
 
 add_unit_test(InterpolationCompanion
     mechanics/interpolation/InterpolationTrussLinear.cpp
     mechanics/interpolation/InterpolationTrussQuadratic.cpp
+    mechanics/interpolation/InterpolationTrussLobatto.cpp
     mechanics/interpolation/InterpolationTriangleLinear.cpp
     mechanics/interpolation/InterpolationTriangleQuadratic.cpp
     mechanics/interpolation/InterpolationQuadLinear.cpp
     mechanics/interpolation/InterpolationQuadSerendipity.cpp
+    mechanics/interpolation/InterpolationQuadLobatto.cpp
     mechanics/interpolation/InterpolationTetrahedronLinear.cpp
     mechanics/interpolation/InterpolationTetrahedronQuadratic.cpp
     mechanics/interpolation/InterpolationBrickLinear.cpp
+    mechanics/interpolation/InterpolationBrickLobatto.cpp
     mechanics/interpolation/InterpolationPrismLinear.cpp
     mechanics/interpolation/InterpolationPrismQuadratic.cpp
     mechanics/interpolation/InterpolationPyramidLinear.cpp
